+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ### Added
 
+- `PagedListLayout` now supports a peek value, which allows items to peek into view from the leading/top and trailing/bottom edges.
+- Replacing the `isPagingEnabled` boolean with a `PagingStyle` enum. This gives layouts the ability to:
+  - leverage `native` UIScrollView paging, where items are full width (equivalent to `isPagingEnabled` being `true`)
+  - use `custom` paging, where items can be any width (this is new functionality)
+  - disable paging (equivalent to `isPagingEnabled` being `false`)
+- Adding a `percentageVisible` property to `ListScrollPositionInfo.VisibleItem`.
+- Adding a `mostVisibleItem` property to `ListScrollPositionInfo`.
+
 ### Removed
 
 ### Changed
@@ -11,6 +19,8 @@
 ### Misc
 
 ### Internal
+
+- Replacing `PagedAppearance.PagingSize.view` with a `.inset(Peek)` case. This is used by `PagedListLayout` to lay out items wiht an edge peek.
 
 # Past Releases