+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ### Internal
 
 - Replacing `PagedAppearance.PagingSize.view` with a `.inset(Peek)` case. This is used by `PagedListLayout` to lay out items wiht an edge peek.
+- Updating GitHub `actions/cache` to v4.
 
 # Past Releases